+++ smarts/core/vehicle.py
@@ -569,7 +569,7 @@ class Vehicle:
                 detach = getattr(self, f"detach_{sensor_name}")
                 if detach:
                     detach(sensor_name)
-                    self._log.info(
+                    self._log.debug(
                         f"replacing existing {sensor_name} on vehicle {self.id}"
                     )
                 setattr(self, f"_{sensor_name}", sensor)